# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductsController_createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid category or file references",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new product listing",
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_getProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details with images",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a product by ID",
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Not the product owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a product listing",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_deleteProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted (soft)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDeleteResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Not the product owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a product listing",
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/status": {
      "patch": {
        "operationId": "ProductsController_updateProductStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductStatusResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Not the product owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product status (available/sold/archived)",
        "tags": [
          "Products"
        ]
      }
    },
    "/my/products": {
      "get": {
        "operationId": "ProductsController_listMyProducts",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category enum key",
            "schema": {
              "example": "electronics",
              "type": "string",
              "enum": [
                "realEstate",
                "vehicles",
                "electronics",
                "homeAndDecor",
                "clothingAndFashion",
                "services"
              ]
            }
          },
          {
            "name": "subcategory",
            "required": false,
            "in": "query",
            "description": "Filter by subcategory enum key (use all as wildcard)",
            "schema": {
              "example": "all",
              "type": "string",
              "enum": [
                "all",
                "apartmentForRent",
                "apartmentForSale",
                "houses",
                "lands",
                "commercialRealEstateForRent",
                "commercialRealEstateForSale",
                "other",
                "carsForSale",
                "carsForRent",
                "sparePartsAndAccessories",
                "motorCycles",
                "bicycles",
                "trucksAndHeavyVehicles",
                "smartphones",
                "tablets",
                "laptopsAndComputers",
                "accessories",
                "speakersAndHeadphones",
                "cameras",
                "smartWatchesAndWearables",
                "monitorsAndTVs",
                "furniture",
                "officeFurniture",
                "kitchenAndDining",
                "beddingAndBath",
                "homeDecor",
                "homeTools",
                "lighting",
                "menClothing",
                "womenClothing",
                "kidsClothing",
                "shoes",
                "menAccessories",
                "womenAccessoriesAndMakeup",
                "jewelryAndWatches",
                "maintenanceAndRepairs",
                "transportationAndMoving",
                "personalServices",
                "carsServices",
                "homeServices",
                "lessonsAndTutoring"
              ]
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum price filter",
            "schema": {
              "minimum": 0,
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price filter",
            "schema": {
              "minimum": 0,
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "description": "Filter products listed on or after this date (ISO 8601)",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "description": "Filter products listed on or before this date (ISO 8601); must be ≥ fromDate",
            "schema": {
              "example": "2024-12-31",
              "type": "string"
            }
          },
          {
            "name": "minRate",
            "required": false,
            "in": "query",
            "description": "Minimum seller average rating (0–5)",
            "schema": {
              "minimum": 0,
              "maximum": 5,
              "example": 3.5,
              "type": "number"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "description": "Filter by city",
            "schema": {
              "example": "Cairo",
              "type": "string"
            }
          },
          {
            "name": "addressText",
            "required": false,
            "in": "query",
            "description": "Filter by address text (partial match)",
            "schema": {
              "example": "Tahrir",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Full-text search across product name and description",
            "schema": {
              "example": "iPhone",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "example": "created",
              "type": "string",
              "enum": [
                "price",
                "address",
                "rate",
                "created"
              ]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "description": "Sort direction",
            "schema": {
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (default 0)",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by product status",
            "schema": {
              "example": "available",
              "type": "string",
              "enum": [
                "available",
                "sold",
                "archived"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of own products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the current user's products",
        "tags": [
          "Products"
        ]
      }
    },
    "/search/products": {
      "get": {
        "operationId": "ProductsController_searchProducts",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category enum key",
            "schema": {
              "example": "electronics",
              "type": "string",
              "enum": [
                "realEstate",
                "vehicles",
                "electronics",
                "homeAndDecor",
                "clothingAndFashion",
                "services"
              ]
            }
          },
          {
            "name": "subcategory",
            "required": false,
            "in": "query",
            "description": "Filter by subcategory enum key (use all as wildcard)",
            "schema": {
              "example": "all",
              "type": "string",
              "enum": [
                "all",
                "apartmentForRent",
                "apartmentForSale",
                "houses",
                "lands",
                "commercialRealEstateForRent",
                "commercialRealEstateForSale",
                "other",
                "carsForSale",
                "carsForRent",
                "sparePartsAndAccessories",
                "motorCycles",
                "bicycles",
                "trucksAndHeavyVehicles",
                "smartphones",
                "tablets",
                "laptopsAndComputers",
                "accessories",
                "speakersAndHeadphones",
                "cameras",
                "smartWatchesAndWearables",
                "monitorsAndTVs",
                "furniture",
                "officeFurniture",
                "kitchenAndDining",
                "beddingAndBath",
                "homeDecor",
                "homeTools",
                "lighting",
                "menClothing",
                "womenClothing",
                "kidsClothing",
                "shoes",
                "menAccessories",
                "womenAccessoriesAndMakeup",
                "jewelryAndWatches",
                "maintenanceAndRepairs",
                "transportationAndMoving",
                "personalServices",
                "carsServices",
                "homeServices",
                "lessonsAndTutoring"
              ]
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum price filter",
            "schema": {
              "minimum": 0,
              "example": 100,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price filter",
            "schema": {
              "minimum": 0,
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "description": "Filter products listed on or after this date (ISO 8601)",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "description": "Filter products listed on or before this date (ISO 8601); must be ≥ fromDate",
            "schema": {
              "example": "2024-12-31",
              "type": "string"
            }
          },
          {
            "name": "minRate",
            "required": false,
            "in": "query",
            "description": "Minimum seller average rating (0–5)",
            "schema": {
              "minimum": 0,
              "maximum": 5,
              "example": 3.5,
              "type": "number"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "description": "Filter by city",
            "schema": {
              "example": "Cairo",
              "type": "string"
            }
          },
          {
            "name": "addressText",
            "required": false,
            "in": "query",
            "description": "Filter by address text (partial match)",
            "schema": {
              "example": "Tahrir",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Full-text search across product name and description",
            "schema": {
              "example": "iPhone",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "example": "created",
              "type": "string",
              "enum": [
                "price",
                "address",
                "rate",
                "created"
              ]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "description": "Sort direction",
            "schema": {
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (default 0)",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Search / filter public product listings",
        "tags": [
          "Products"
        ]
      }
    },
...
    "/v2/products/{id}": {
      "get": {
        "operationId": "KayanController_getProduct",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KayanProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get Kayan product by id",
        "tags": [
          "Kayan V2"
        ]
      }
    },
...
  },
  "info": {
    "title": "Market Place API",
    "description": "Online marketplace REST API — user registration, product listings, real-time chat, ratings, and admin moderation.\n\n> **WebSocket (Chat):** Real-time messaging uses a Socket.io gateway at the `/chat` namespace. That interface is not documented here; see the project README for event names and payloads.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "data": {
            "type": "object",
            "additionalProperties": false,
            "example": null,
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "error"
        ]
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "description": "Product category enum key",
            "enum": [
              "realEstate",
              "vehicles",
              "electronics",
              "homeAndDecor",
              "clothingAndFashion",
              "services"
            ],
            "example": "electronics"
          },
          "subcategory": {
            "type": "string",
            "description": "Product subcategory enum key",
            "enum": [
              "all",
              "apartmentForRent",
              "apartmentForSale",
              "houses",
              "lands",
              "commercialRealEstateForRent",
              "commercialRealEstateForSale",
              "other",
              "carsForSale",
              "carsForRent",
              "sparePartsAndAccessories",
              "motorCycles",
              "bicycles",
              "trucksAndHeavyVehicles",
              "smartphones",
              "tablets",
              "laptopsAndComputers",
              "accessories",
              "speakersAndHeadphones",
              "cameras",
              "smartWatchesAndWearables",
              "monitorsAndTVs",
              "furniture",
              "officeFurniture",
              "kitchenAndDining",
              "beddingAndBath",
              "homeDecor",
              "homeTools",
              "lighting",
              "menClothing",
              "womenClothing",
              "kidsClothing",
              "shoes",
              "menAccessories",
              "womenAccessoriesAndMakeup",
              "jewelryAndWatches",
              "maintenanceAndRepairs",
              "transportationAndMoving",
              "personalServices",
              "carsServices",
              "homeServices",
              "lessonsAndTutoring"
            ],
            "example": "smartphones"
          },
          "name": {
            "type": "string",
            "description": "Product title (1–255 chars)",
            "example": "iPhone 14 Pro Max",
            "minLength": 1,
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Product description (1–5000 chars)",
            "example": "Excellent condition, barely used.",
            "minLength": 1,
            "maxLength": 5000
          },
          "price": {
            "type": "number",
            "description": "Price in the local currency",
            "example": 1500,
            "minimum": 0,
            "maximum": 9999999999.99
          },
          "city": {
            "type": "string",
            "description": "City where the product is located",
            "example": "Cairo",
            "minLength": 1,
            "maxLength": 255
          },
          "addressText": {
            "type": "string",
            "description": "Street / area address text (1–1000 chars)",
            "example": "15 Tahrir Square, Downtown",
            "minLength": 1,
            "maxLength": 1000
          },
          "details": {
            "type": "object",
            "description": "Additional product details as a JSON object",
            "example": {
              "condition": "used",
              "color": "black",
              "storage": "256GB"
            }
          },
          "imageFileIds": {
            "description": "Up to 10 pre-uploaded file IDs for product images",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "isNegotiable": {
            "type": "boolean",
            "description": "Whether price is negotiable",
            "example": true
          },
          "preferredContactMethod": {
            "type": "string",
            "description": "Preferred contact method",
            "enum": [
              "phone",
              "chat",
              "both"
            ],
            "example": "both"
          }
        },
        "required": [
          "category",
          "name",
          "description",
          "price",
          "city",
          "addressText"
        ]
      },
      "ProductDataDto": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/ProductDto"
          }
        },
        "required": [
          "product"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ProductDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "ProductDeleteResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ProductDeleteDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "UpdateProductStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "available",
              "sold",
              "archived"
            ],
            "description": "New product status",
            "example": "sold"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "ProductStatusResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ProductStatusDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "ProductListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ProductListDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "KayanProductResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/KayanProductDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
    }
  }
}